# docs/contracts/openapi/opm-local-api-v1.yaml
openapi: 3.1.0
info:
  title: OPM Local Modeling API
  version: 0.2.0-draft
  summary: P01-P03 与最小版本、校验闭环的本地 HTTP 契约
  description: >-
    该契约仅用于首批开发。服务只监听 loopback；写请求除本地会话头外，
    运行时还必须校验 Host 与 Origin。契约不表示完整 ISO 19450:2024 符合性。
servers:
  - url: http://127.0.0.1:{port}/api/v1
    description: 本地运行时
    variables:
      port:
        default: "17850"
security:
  - localSession: []
...
paths:
  /projects:
    get:
      tags: [Projects]
      operationId: API-PRJ-001
      summary: 分页查询本地项目
      parameters:
        - $ref: "#/components/parameters/Cursor"
        - $ref: "#/components/parameters/PageSize"
        - name: query
          in: query
          schema: {type: string, maxLength: 200}
      responses:
        "200":
          description: 项目页
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectPageResult"
        default:
          $ref: "#/components/responses/ErrorResponse"
    post:
      tags: [Projects]
      operationId: API-PRJ-003
      summary: 创建本地项目
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProjectRequest"
      responses:
        "201":
          description: 项目已创建
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectCommandResult"
        "409":
          $ref: "#/components/responses/ErrorResponse"
        default:
          $ref: "#/components/responses/ErrorResponse"
  /projects/{projectId}:
    parameters:
      - $ref: "#/components/parameters/ProjectId"
    get:
      tags: [Projects]
      operationId: API-PRJ-002
      summary: 读取项目摘要
      responses:
        "200":
          description: 项目摘要
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectQueryResult"
        "404":
          $ref: "#/components/responses/ErrorResponse"
        default:
          $ref: "#/components/responses/ErrorResponse"
  /projects/{projectId}/models:
    parameters:
      - $ref: "#/components/parameters/ProjectId"
    get:
      tags: [Models]
      operationId: API-PRJ-006
      summary: 分页查询项目中的模型
      parameters:
        - $ref: "#/components/parameters/Cursor"
        - $ref: "#/components/parameters/PageSize"
      responses:
        "200":
          description: 模型页
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ModelPageResult"
        default:
          $ref: "#/components/responses/ErrorResponse"
    post:
      tags: [Models]
      operationId: API-PRJ-007
      summary: 创建模型、初始草稿和根 Context
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateModelRequest"
      responses:
        "201":
          description: 模型已创建
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ModelCommandResult"
        "409":
          $ref: "#/components/responses/ErrorResponse"
        default:
          $ref: "#/components/responses/ErrorResponse"
  /projects/{projectId}/models/{modelId}/workspace-session:
    parameters:
      - $ref: "#/components/parameters/ProjectId"
      - $ref: "#/components/parameters/ModelId"
    get:
      tags: [Models]
      operationId: API-PRJ-009
      summary: 打开模型并取得工作台会话
      description: 省略 revision 或传 head 解析活动草稿；无草稿时解析最近 Baseline。精确 Revision 校验 Model 归属并始终只读。Context 在目标版本解析，缺失或非法时返回目标根；无可用版本返回 404。
      parameters:
        - name: revision
          in: query
          required: false
          schema: {$ref: "#/components/schemas/StableId"}
        - name: context
          in: query
          required: false
          schema: {type: string}
      responses:
        "200":
          description: 工作台会话
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkspaceSessionResult"
        "404":
          $ref: "#/components/responses/ErrorResponse"
        default:
          $ref: "#/components/responses/ErrorResponse"
  /projects/{projectId}/models/{modelId}/contexts/{contextId}/navigation:
    parameters:
      - $ref: "#/components/parameters/ProjectId"
      - $ref: "#/components/parameters/ModelId"
      - $ref: "#/components/parameters/ContextId"
      - $ref: "#/components/parameters/Revision"
    get:
      tags: [Contexts]
      operationId: API-CTX-001
      summary: 读取多 OPD 导航树和当前 Context 路径
      responses:
        "200":
          description: Context 导航投影
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContextNavigationResult"
        default:
          $ref: "#/components/responses/ErrorResponse"
  /projects/{projectId}/models/{modelId}/contexts/{contextId}/projection:
    parameters:
      - $ref: "#/components/parameters/ProjectId"
      - $ref: "#/components/parameters/ModelId"
      - $ref: "#/components/parameters/ContextId"
      - $ref: "#/components/parameters/Revision"
    get:
      tags: [Contexts]
      operationId: API-CTX-002
      summary: 读取 OPD Construct 投影
      responses:
        "200":
          description: 与修订绑定的 OPD 投影
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContextProjectionResult"
        default:
          $ref: "#/components/responses/ErrorResponse"
  /projects/{projectId}/models/{modelId}/contexts/{contextId}/command-capabilities:
    parameters:
      - $ref: "#/components/parameters/ProjectId"
      - $ref: "#/components/parameters/ModelId"
      - $ref: "#/components/parameters/ContextId"
      - $ref: "#/components/parameters/Revision"
    get:
      tags: [Editing]
      operationId: API-EDT-001
      summary: 查询当前选择可执行的建模命令
      parameters:
        - name: selection_id
          in: query
          schema: {type: string}
        - name: intent
          in: query
          description: 完整画布候选查询意图；未提供时保持 P0 命令目录行为。
          schema:
            type: string
            enum: [CREATE_FEATURE, CREATE_STATE, CREATE_FACT, UPDATE_STATE, UPDATE_FACT, DELETE_CONSTRUCT]
        - name: endpoint
          in: query
          description: 用户选择顺序的 endpoint locator，可重复；服务端返回规范化端点。
          style: form
          explode: true
          schema:
            type: array
            items: {$ref: "#/components/schemas/StableId"}
      responses:
        "200":
          description: Profile 与 Rule 过滤后的命令能力
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CommandCapabilitiesResult"
        default:
          $ref: "#/components/responses/ErrorResponse"
  /projects/{projectId}/models/{modelId}/contexts/{contextId}/relation-catalog:
    parameters:
      - $ref: "#/components/parameters/ProjectId"
      - $ref: "#/components/parameters/ModelId"
      - $ref: "#/components/parameters/ContextId"
      - $ref: "#/components/parameters/Revision"
    get:
      tags: [Catalog]
      operationId: API-CAT-001
      summary: 读取活动 Profile 的全量关系工具目录
      description: >-
        目录只表示可见 capability，不是创建或更新 Fact 的授权。Control 条目只能通过既有 Procedural Fact 的 UPDATE_FACT 候选形成。
      parameters:
        - name: selection_id
          in: query
          description: 当前 Context 内选中的 committed Fact；仅用于计算 Control 目录项可用性。
          schema: {type: string}
      responses:
        "200":
          description: 与修订绑定的关系目录
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RelationCatalogResult"
        default:
          $ref: "#/components/responses/ErrorResponse"
  /projects/{projectId}/models/{modelId}/contexts/{contextId}/release-visual-common-fault-command:
    parameters:
      - $ref: "#/components/parameters/ProjectId"
      - $ref: "#/components/parameters/ModelId"
      - $ref: "#/components/parameters/ContextId"
      - $ref: "#/components/parameters/Revision"
    get:
      tags: [ReleaseAuthoring]
      operationId: API-REL-003
      summary: 发现受控 Visual Common 一次性故障命令
      description: >-
        仅在 release-only Runtime 已通过 one-shot fault launch guard 且请求 identity 完全匹配 BLOCKED_FEEDBACK fixture 时返回固定 descriptor；普通 Runtime 返回 NOT_FOUND。
      responses:
        "200":
          description: 受控命令 descriptor
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReleaseVisualCommonFaultCommandResult"
        default:
          $ref: "#/components/responses/ErrorResponse"
  /projects/{projectId}/models/{modelId}/contexts/{contextId}/commands:
    parameters:
      - $ref: "#/components/parameters/ProjectId"
      - $ref: "#/components/parameters/ModelId"
      - $ref: "#/components/parameters/ContextId"
    post:
      tags: [Editing]
      operationId: API-EDT-002
      summary: 原子执行一个语义编辑命令
      description: >-
        普通视口缩放、平移和面板状态禁止调用本接口。语义 in/out-zoom、
        Context 变更和语义布局必须调用本接口并产生新修订。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExecuteEditCommandRequest"
      responses:
        "200":
          description: 命令已原子提交或被业务规则阻断
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EditCommandResult"
        "409":
          $ref: "#/components/responses/ErrorResponse"
        "422":
          $ref: "#/components/responses/ErrorResponse"
        default:
          $ref: "#/components/responses/ErrorResponse"
  /projects/{projectId}/models/{modelId}/contexts/{contextId}/text-projection:
    parameters:
      - $ref: "#/components/parameters/ProjectId"
      - $ref: "#/components/parameters/ModelId"
      - $ref: "#/components/parameters/ContextId"
      - $ref: "#/components/parameters/Revision"
    get:
      tags: [Text]
      operationId: API-TXT-001
      summary: 读取与当前 OPD 修订一致的 OPL 或 OPT 投影
      parameters:
        - name: scope
          in: query
          schema:
            type: string
            enum: [CONTEXT, SUBTREE, MODEL]
            default: CONTEXT
      responses:
        "200":
          description: 文本投影和图文追踪索引
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TextProjectionResult"
        default:
          $ref: "#/components/responses/ErrorResponse"
  /projects/{projectId}/models/{modelId}/contexts/{contextId}/findings:
    parameters:
      - $ref: "#/components/parameters/ProjectId"
      - $ref: "#/components/parameters/ModelId"
      - $ref: "#/components/parameters/ContextId"
      - $ref: "#/components/parameters/Revision"
    get:
      tags: [Validation]
      operationId: API-VAL-003
      summary: 读取 Context 内与修订绑定的 Finding 列表
      responses:
        "200":
          description: Finding 列表
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FindingListResult"
        default:
          $ref: "#/components/responses/ErrorResponse"
  /projects/{projectId}/models/{modelId}/contexts/{contextId}/operation-records:
    parameters:
      - $ref: "#/components/parameters/ProjectId"
      - $ref: "#/components/parameters/ModelId"
      - $ref: "#/components/parameters/ContextId"
      - $ref: "#/components/parameters/Revision"
    get:
      tags: [Versions]
      operationId: API-VER-006
      summary: 读取与修订绑定的 Operation Record 历史
      responses:
        "200":
          description: Operation Record 列表
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OperationRecordListResult"
        default:
          $ref: "#/components/responses/ErrorResponse"
  /projects/{projectId}/models/{modelId}/validation-tasks:
    parameters:
      - $ref: "#/components/parameters/ProjectId"
      - $ref: "#/components/parameters/ModelId"
    post:
      tags: [Validation]
      operationId: API-VAL-001
      summary: 对固定输入修订启动校验任务
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ValidateModelRequest"
      responses:
        "202":
          description: 校验任务已耐久登记
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskAcceptedResult"
        "409":
          $ref: "#/components/responses/ErrorResponse"
        default:
          $ref: "#/components/responses/ErrorResponse"
  /projects/{projectId}/models/{modelId}/revisions:
    parameters:
      - $ref: "#/components/parameters/ProjectId"
      - $ref: "#/components/parameters/ModelId"
    get:
      tags: [Versions]
      operationId: API-VER-001
      summary: 查询草稿、快照和基线修订
      parameters:
        - $ref: "#/components/parameters/Cursor"
        - $ref: "#/components/parameters/PageSize"
      responses:
        "200":
          description: 修订页
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RevisionPageResult"
        default:
          $ref: "#/components/responses/ErrorResponse"
  /projects/{projectId}/models/{modelId}/baselines:
    parameters:
      - $ref: "#/components/parameters/ProjectId"
      - $ref: "#/components/parameters/ModelId"
    post:
      tags: [Versions]
      operationId: API-VER-004
      summary: 在全量校验门槛后创建不可变基线
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBaselineRequest"
      responses:
        "201":
          description: 基线已创建
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BaselineCommandResult"
        "202":
          description: 基线门槛校验任务已耐久登记
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskAcceptedResult"
        "409":
          $ref: "#/components/responses/ErrorResponse"
        "422":
          $ref: "#/components/responses/ErrorResponse"
        default:
          $ref: "#/components/responses/ErrorResponse"
  /tasks/{taskId}:
    parameters:
      - $ref: "#/components/parameters/TaskId"
    get:
      tags: [Tasks]
      operationId: API-TSK-001
      summary: 查询后台任务状态和结果
      responses:
        "200":
          description: 任务状态
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskQueryResult"
        "404":
          $ref: "#/components/responses/ErrorResponse"
        default:
          $ref: "#/components/responses/ErrorResponse"
  /tasks/{taskId}/events:
    parameters:
      - $ref: "#/components/parameters/TaskId"
    get:
      tags: [Tasks]
      operationId: API-TSK-004
      summary: 订阅单个任务的本地 SSE 状态流
      responses:
        "200":
          description: 任务事件流；断线后客户端必须回查 API-TSK-001
          content:
            text/event-stream:
              schema:
                type: string
                examples:
                  - "event: task-progress\ndata: {\"task_id\":\"task.validate.001\",\"state\":\"RUNNING\",\"progress\":40}\n\n"
        default:
          $ref: "#/components/responses/ErrorResponse"
components:
  securitySchemes:
    localSession:
      type: apiKey
      in: header
      name: X-OPM-Session
      description: 启动本地运行时后签发的短期会话值，不表示用户身份或权限。
  parameters:
    ProjectId:
      name: projectId
      in: path
      required: true
      schema: {$ref: "#/components/schemas/StableId"}
    ModelId:
      name: modelId
      in: path
      required: true
      schema: {$ref: "#/components/schemas/StableId"}
    ContextId:
      name: contextId
      in: path
      required: true
      schema: {$ref: "#/components/schemas/StableId"}
...
    Revision:
      name: revision
      in: query
      required: true
      schema: {$ref: "#/components/schemas/StableId"}
    Cursor:
      name: cursor
      in: query
      schema: {type: string}
    PageSize:
      name: page_size
      in: query
      schema: {type: integer, minimum: 1, maximum: 200, default: 50}
  responses:
    ErrorResponse:
      description: 结构化错误；具体 HTTP 状态由错误类别决定
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
  schemas:
    StableId:
      type: string
      minLength: 3
      maxLength: 160
      pattern: "^[A-Za-z][A-Za-z0-9._:-]*$"
...
    ProjectPageResult:
      type: object
      required: [meta, data]
      properties:
        meta: {$ref: "#/components/schemas/QueryMeta"}
        data: {type: array, items: {$ref: "#/components/schemas/Project"}}
    ProjectQueryResult:
      type: object
      required: [meta, data]
      properties:
        meta: {$ref: "#/components/schemas/QueryMeta"}
        data: {$ref: "#/components/schemas/Project"}
    ModelPageResult:
      type: object
      required: [meta, data]
      properties:
        meta: {$ref: "#/components/schemas/QueryMeta"}
        data: {type: array, items: {$ref: "#/components/schemas/Model"}}
...
    CreateProjectRequest:
      type: object
      unevaluatedProperties: false
      allOf:
        - $ref: "#/components/schemas/WriteGuard"
        - type: object
          required: [name]
          properties:
            name: {type: string, minLength: 1, maxLength: 256}
            description: {type: string, maxLength: 4000}
    CreateModelRequest:
      type: object
      unevaluatedProperties: false
      allOf:
        - $ref: "#/components/schemas/WriteGuard"
        - type: object
          required: [name, binding]
          properties:
            name: {type: string, minLength: 1, maxLength: 256}
            description: {type: string, maxLength: 4000}
            binding: {$ref: "#/components/schemas/ProfileRuleGuard"}
    ProjectCommandResult:
      type: object
      required: [meta, data]
      properties:
        meta: {$ref: "#/components/schemas/CommandMeta"}
        data: {$ref: "#/components/schemas/Project"}
    ModelCommandResult:
      type: object
      required: [meta, data]
      properties:
        meta: {$ref: "#/components/schemas/CommandMeta"}
        data: {$ref: "#/components/schemas/Model"}
    WorkspaceSessionResult:
      type: object
      required: [meta, data]
      properties:
        meta: {$ref: "#/components/schemas/QueryMeta"}
        data:
          type: object
          required: [model, root_context_id, current_context_id, viewport_state, panel_state]
          properties:
            model: {$ref: "#/components/schemas/Model"}
            root_context_id: {$ref: "#/components/schemas/StableId"}
            current_context_id: {$ref: "#/components/schemas/StableId"}
            viewport_state: {type: object, additionalProperties: true}
            panel_state: {type: object, additionalProperties: true}
    ContextNavigationResult:
      type: object
      required: [meta, data]
      properties:
        meta: {$ref: "#/components/schemas/QueryMeta"}
        data:
          type: object
          required: [current_path, process_tree, object_forest, views]
          properties:
            current_path: {type: array, items: {$ref: "#/components/schemas/StableId"}}
            process_tree: {type: array, items: {$ref: "#/components/schemas/NavigationNode"}}
            object_forest: {type: array, items: {$ref: "#/components/schemas/NavigationNode"}}
            views: {type: array, items: {$ref: "#/components/schemas/NavigationNode"}}
...
    ContextProjectionResult:
      type: object
      required: [meta, data]
      properties:
        meta: {$ref: "#/components/schemas/QueryMeta"}
        data:
          type: object
          required: [context_id, constructs, suppressed_states]
          properties:
            context_id: {$ref: "#/components/schemas/StableId"}
            constructs:
              type: array
              items:
                type: object
                required: [occurrence_id, target_id, target_kind, construct_role, capability_id, layout]
                properties:
                  occurrence_id: {$ref: "#/components/schemas/StableId"}
                  target_id: {$ref: "#/components/schemas/StableId"}
                  target_kind: {type: string, enum: [ELEMENT, FEATURE, STATE, FACT]}
                  construct_role: {type: string}
                  label: {type: string}
                  owner_id: {$ref: "#/components/schemas/StableId"}
                  state_roles: {type: array, items: {type: string, enum: [INITIAL, DEFAULT, FINAL]}}
                  explicitness: {type: string, enum: [EXPLICIT, SUPPRESSED]}
                  fold_state: {type: string, enum: [UNFOLDED, FOLDED]}
                  direction: {type: string, enum: [DIRECTED, BIDIRECTIONAL, UNDIRECTED]}
                  capability_id: {type: string}
                  endpoints: {type: array, items: {type: object, additionalProperties: true}}
                  modifiers: {type: array, items: {type: object, additionalProperties: true}}
                  labels: {type: array, items: {type: object, additionalProperties: true}}
                  collection_completeness: {type: string, enum: [COMPLETE, INCOMPLETE, NOT_APPLICABLE]}
                  layout: {type: object, additionalProperties: true}
            suppressed_states:
              type: array
              items:
                $ref: "#/components/schemas/SuppressedState"
...
    CommandCapabilitiesResult:
      type: object
      additionalProperties: false
      required: [meta, data]
      properties:
        meta: {$ref: "#/components/schemas/QueryMeta"}
        data:
          type: object
          additionalProperties: false
          required: [allowed, forbidden, capability_query_id, options]
          properties:
            allowed:
              type: array
              items: {$ref: "#/components/schemas/EditCommandType"}
            forbidden:
              type: array
              items:
                $ref: "#/components/schemas/ForbiddenCommand"
            capability_query_id: {$ref: "#/components/schemas/StableId"}
            options:
              type: array
              items: {$ref: "#/components/schemas/CommandCapabilityOption"}
    RelationCatalogResult:
      type: object
      additionalProperties: false
      required: [meta, data]
      properties:
        meta: {$ref: "#/components/schemas/QueryMeta"}
        data:
          type: object
          additionalProperties: false
          required: [items]
          properties:
            items:
              type: array
              minItems: 34
              maxItems: 34
              items: {$ref: "#/components/schemas/RelationCatalogItem"}
...
    ReleaseVisualCommonFaultCommandResult:
      type: object
      additionalProperties: false
      required: [meta, data]
      properties:
        meta: {$ref: "#/components/schemas/QueryMeta"}
        data:
          type: object
          additionalProperties: false
          required: [command_id, command_type, payload]
          properties:
            command_id: {const: command.visual.blocked-feedback.persistence-failed}
            command_type: {const: CREATE_FACT}
            payload:
              type: object
              additionalProperties: false
              required: [kind, fact_id, object_id, process_id, layout]
              properties:
                kind: {const: CONSUMPTION}
                fact_id: {const: fact.visual.blocked-feedback.one-shot}
                object_id: {const: element.visual.blocked-feedback.input}
                process_id: {const: element.visual.blocked-feedback.process}
                layout:
                  type: object
                  additionalProperties: false
                  required: [x, y]
                  properties:
                    x: {const: 340}
                    y: {const: 266}
    ExecuteEditCommandRequest:
      type: object
      unevaluatedProperties: false
      allOf:
        - $ref: "#/components/schemas/WriteGuard"
        - type: object
          required: [base_revision, binding, command_type, payload]
          properties:
            base_revision: {$ref: "#/components/schemas/StableId"}
            binding: {$ref: "#/components/schemas/ProfileRuleGuard"}
            command_type: {$ref: "#/components/schemas/EditCommandType"}
        - oneOf:
            - $ref: "#/components/schemas/CreateElementCommand"
            - $ref: "#/components/schemas/CreateFeatureCommand"
            - $ref: "#/components/schemas/CreateFactCommand"
            - $ref: "#/components/schemas/CreateStateCommand"
            - $ref: "#/components/schemas/UpdateStateCommand"
            - $ref: "#/components/schemas/UpdateFactCommand"
            - $ref: "#/components/schemas/UpdatePropertyCommand"
            - $ref: "#/components/schemas/UpdateLayoutCommand"
            - $ref: "#/components/schemas/DeleteConstructCommand"
            - $ref: "#/components/schemas/StateExplicitCommand"
            - $ref: "#/components/schemas/StateSuppressCommand"
            - $ref: "#/components/schemas/UnfoldCommand"
            - $ref: "#/components/schemas/FoldCommand"
            - $ref: "#/components/schemas/LegacyEditCommand"
...
    EditCommandResult:
      type: object
      required: [meta, data]
      properties:
        meta: {$ref: "#/components/schemas/CommandMeta"}
        data:
          type: object
          required: [affected_ids, text_trace_ids, validation_summary]
          properties:
            affected_ids: {type: array, items: {$ref: "#/components/schemas/StableId"}}
            text_trace_ids: {type: array, items: {$ref: "#/components/schemas/StableId"}}
            validation_summary: {$ref: "#/components/schemas/ValidationSummary"}
...
    FindingListResult:
      type: object
      additionalProperties: false
      required: [meta, data]
      properties:
        meta: {$ref: "#/components/schemas/QueryMeta"}
        data:
          type: array
          items:
            type: object
            additionalProperties: false
            required: [finding_id, rule_id, severity, category, context_id, entity_id]
            properties:
              finding_id: {$ref: "#/components/schemas/StableId"}
              rule_id: {$ref: "#/components/schemas/StableId"}
              severity: {type: string, enum: [BLOCKING, WARNING, SUGGESTION]}
              category: {type: string, minLength: 1}
              context_id: {$ref: "#/components/schemas/StableId"}
              entity_id: {$ref: "#/components/schemas/StableId"}
...
    TaskAcceptedResult:
      type: object
      required: [meta, data]
      properties:
        meta: {$ref: "#/components/schemas/CommandMeta"}
        data: {$ref: "#/components/schemas/TaskDescriptor"}
...